testing... 1, 2, 3...

+++ _replace.sh
@@ -10,6 +10,12 @@ name='xxxNAMExxx'
 cname='xxxCNAMExxx'
 
 mv $name $1
+
+if [ $? -ne 0 ]
+        then
+                exit
+fi
+
 rm README.md
 mv xxxREADME.md README.md
 
@@ -20,7 +26,8 @@ mv $cname.pxd $2.pxd
 
 popd
 
-
 egrep -lRZ $name . | xargs -0 -l sed -i -e "s/$name/$1/g"
 egrep -lRZ $cname . | xargs -0 -l sed -i -e "s/$cname/$2/g"
 
+# suicide
+rm _replace.sh
